Move EnabledExtensions to Configuration
